ajustes na versao 0.1.0 para ambiente de producao

+++ src/main/java/br/com/altamira/monitoramento/model/IHMLog.java
@@ -40,7 +40,7 @@ public class IHMLog {
 	private long sequencia;
 	
 	@Column(name = "TEMPO")
-	private int tempo;
+	private long tempo;
 	
 	@Column(name = "OPERADOR")
 	private String operador;
@@ -50,7 +50,7 @@ public class IHMLog {
 	}
 
 	public IHMLog(String ihm, Date datahora, int modo, long sequencia,
-			int tempo, String operador) {
+			long tempo, String operador) {
 		super();
 		this.ihm = ihm;
 		this.datahora = datahora;
@@ -116,11 +116,11 @@ public class IHMLog {
 		this.sequencia = sequencia;
 	}
 
-	public int getTempo() {
+	public long getTempo() {
 		return tempo;
 	}
 
-	public void setTempo(int tempo) {
+	public void setTempo(long tempo) {
 		this.tempo = tempo;
 	}
